fix(types): update inventory to use int64

+++ types/commons.go
@@ -25,13 +25,13 @@ const (
 
 // Inventory stores the items in a vendor's inventory
 type Inventory struct {
-	Truck         int `json:"Truck,omitempty" bson:"Truck,omitempty"`
-	Crane         int `json:"Crane,omitempty" bson:"Crane,omitempty"`
-	Tanker        int `json:"Tanker,omitempty" bson:"Tanker,omitempty"`
-	RoadRoller    int `json:"RoadRoller,omitempty" bson:"RoadRoller,omitempty"`
-	ForkLift      int `json:"ForkLift,omitempty" bson:"ForkLift,omitempty"`
-	BoomLifter    int `json:"BoomLifter,omitempty" bson:"BoomLifter,omitempty"`
-	ManLifter     int `json:"ManLifter,omitempty" bson:"ManLifter,omitempty"`
-	HydraulicJack int `json:"HydraulicJack,omitempty" bson:"HydraulicJack,omitempty"`
-	Manpower      int `json:"Manpower,omitempty" bson:"Manpower,omitempty"`
+	Truck         int64 `json:"Truck,omitempty" bson:"Truck,omitempty"`
+	Crane         int64 `json:"Crane,omitempty" bson:"Crane,omitempty"`
+	Tanker        int64 `json:"Tanker,omitempty" bson:"Tanker,omitempty"`
+	RoadRoller    int64 `json:"RoadRoller,omitempty" bson:"RoadRoller,omitempty"`
+	ForkLift      int64 `json:"ForkLift,omitempty" bson:"ForkLift,omitempty"`
+	BoomLifter    int64 `json:"BoomLifter,omitempty" bson:"BoomLifter,omitempty"`
+	ManLifter     int64 `json:"ManLifter,omitempty" bson:"ManLifter,omitempty"`
+	HydraulicJack int64 `json:"HydraulicJack,omitempty" bson:"HydraulicJack,omitempty"`
+	Manpower      int64 `json:"Manpower,omitempty" bson:"Manpower,omitempty"`
 }
